+++ query/src/func/window/internal.rs
